Addressed issues with storing the trained KMeans model by making the following changes:
Removed the 'plotters.py' file.
Added plotting functions as methods of the 'KMModel' class.

Additionally, the following enhancements were made:

Implemented the silhouette method for calculating the best number of clusters, enabling automatic selection of the number of clusters for the KMeans model (with the option to choose manually using elbow chart).
Included the ability to plot users locations on other plotted charts.
Added a basic 'main' function."